move enum to file
